+++ tests/test_engine_type_modifiers/__init__.py
@@ -1,0 +1,6 @@
+import os
+
+TEST_CONNECTORS_PATH = os.path.dirname(os.path.abspath(__file__))
+
+# file system path for unittest fixture files
+TEST_DATA_PATH = os.path.join(TEST_CONNECTORS_PATH, "..", "data")